feat: Calculate heart rate and display it

+++ app/src/main/java/com/seoultech/ecgmonitor/heartrate/HeartRateLiveData.kt
@@ -2,9 +2,9 @@ package com.seoultech.ecgmonitor.heartrate
 
 import androidx.lifecycle.LiveData
 
-class HeartRateLiveData: LiveData<Float>() {
+class HeartRateLiveData: LiveData<Int>() {
 
-    fun setHeartRateValue(value: Float) {
+    fun setHeartRateValue(value: Int) {
         postValue(value)
     }
 }
